requirements.txt
aws-cdk-lib>=2.41.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-batch-alpha
aws-cdk.aws_glue_alpha
cdk-nag==2.18.10
botocore
boto3

from-root
requests

pytest
pytest-dotenv
pytest-check
pytest-test-groups
cryptography
moto[all]
aws-requests-auth

pylint
bandit
toml
black